+++ tests/Support/IntTestRouteName.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Router\Tests\Support;
+
+enum IntTestRouteName: int
+{
+    case Home = 1;
+    case About = 2;
+}